+++ packages/dataviews/src/dataviews.tsx
@@ -7,7 +7,7 @@ import type { ComponentType } from 'react';
  * WordPress dependencies
  */
 import { __experimentalHStack as HStack } from '@wordpress/components';
-import { useMemo, useState, useCallback } from '@wordpress/element';
+import { useMemo, useState } from '@wordpress/element';
 
 /**
  * Internal dependencies
@@ -21,9 +21,12 @@ import { VIEW_LAYOUTS } from './layouts';
 import BulkActions from './bulk-actions';
 import { normalizeFields } from './normalize-fields';
 import BulkActionsToolbar from './bulk-actions-toolbar';
-import type { Action, AnyItem, Field, View, ViewBaseProps } from './types';
+import type { Action, Field, View, ViewBaseProps } from './types';
+import type { SetSelection, SelectionOrUpdater } from './private-types';
 
-interface DataViewsProps< Item extends AnyItem > {
+type ItemWithId = { id: string };
+
+type DataViewsProps< Item > = {
 	view: View;
 	onChangeView: ( view: View ) => void;
 	fields: Field< Item >[];
@@ -31,7 +34,6 @@ interface DataViewsProps< Item extends AnyItem > {
 	searchLabel?: string;
 	actions?: Action< Item >[];
 	data: Item[];
-	getItemId?: ( item: Item ) => string;
 	isLoading?: boolean;
 	paginationInfo: {
 		totalItems: number;
@@ -39,14 +41,17 @@ interface DataViewsProps< Item extends AnyItem > {
 	};
 	supportedLayouts: string[];
 	selection?: string[];
-	setSelection?: ( selection: string[] ) => void;
+	setSelection?: SetSelection;
 	onSelectionChange?: ( items: Item[] ) => void;
-}
+} & ( Item extends ItemWithId
+	? { getItemId?: ( item: Item ) => string }
+	: { getItemId: ( item: Item ) => string } );
 
-const defaultGetItemId = ( item: AnyItem ) => item.id;
+const defaultGetItemId = ( item: ItemWithId ) => item.id;
+
 const defaultOnSelectionChange = () => {};
 
-function useSomeItemHasAPossibleBulkAction< Item extends AnyItem >(
+function useSomeItemHasAPossibleBulkAction< Item >(
 	actions: Action< Item >[],
 	data: Item[]
 ) {
@@ -62,7 +67,7 @@ function useSomeItemHasAPossibleBulkAction< Item extends AnyItem >(
 	}, [ actions, data ] );
 }
 
-export default function DataViews< Item extends AnyItem >( {
+export default function DataViews< Item >( {
 	view,
 	onChangeView,
 	fields,
@@ -79,26 +84,22 @@ export default function DataViews< Item extends AnyItem >( {
 	onSelectionChange = defaultOnSelectionChange,
 }: DataViewsProps< Item > ) {
 	const [ selectionState, setSelectionState ] = useState< string[] >( [] );
-	let selection, setSelection;
-	if (
-		selectionProperty !== undefined &&
-		setSelectionProperty !== undefined
-	) {
-		selection = selectionProperty;
-		setSelection = setSelectionProperty;
-	} else {
-		selection = selectionState;
-		setSelection = setSelectionState;
-	}
+	const isUncontrolled =
+		selectionProperty === undefined || setSelectionProperty === undefined;
+	const selection = isUncontrolled ? selectionState : selectionProperty;
+	const setSelection = isUncontrolled
+		? setSelectionState
+		: setSelectionProperty;
 	const [ openedFilter, setOpenedFilter ] = useState< string | null >( null );
 
-	const onSetSelection = useCallback(
-		( items: Item[] ) => {
-			setSelection( items.map( ( item ) => getItemId( item ) ) );
-			onSelectionChange( items );
-		},
-		[ setSelection, getItemId, onSelectionChange ]
-	);
+	function setSelectionWithChange( value: SelectionOrUpdater ) {
+		const newValue =
+			typeof value === 'function' ? value( selection ) : value;
+		onSelectionChange(
+			data.filter( ( item ) => newValue.includes( getItemId( item ) ) )
+		);
+		return setSelection( value );
+	}
 
 	const ViewComponent = VIEW_LAYOUTS.find( ( v ) => v.type === view.type )
 		?.component as ComponentType< ViewBaseProps< Item > >;
@@ -145,7 +146,7 @@ export default function DataViews< Item extends AnyItem >( {
 						<BulkActions
 							actions={ actions }
 							data={ data }
-							onSelectionChange={ onSetSelection }
+							onSelectionChange={ setSelectionWithChange }
 							selection={ _selection }
 							getItemId={ getItemId }
 						/>
@@ -164,7 +165,7 @@ export default function DataViews< Item extends AnyItem >( {
 				getItemId={ getItemId }
 				isLoading={ isLoading }
 				onChangeView={ onChangeView }
-				onSelectionChange={ onSetSelection }
+				onSelectionChange={ setSelectionWithChange }
 				selection={ _selection }
 				setOpenedFilter={ setOpenedFilter }
 				view={ view }
@@ -180,7 +181,7 @@ export default function DataViews< Item extends AnyItem >( {
 						data={ data }
 						actions={ actions }
 						selection={ _selection }
-						onSelectionChange={ onSetSelection }
+						onSelectionChange={ setSelectionWithChange }
 						getItemId={ getItemId }
 					/>
 				) }