+++ packages/editor/src/store/private-actions.js
@@ -370,12 +370,7 @@ export const revertTemplate =
 export const removeTemplates =
 	( items ) =>
 	async ( { registry } ) => {
-		const isResetting = items.every(
-			( item ) =>
-				!! item &&
-				( item.has_theme_file ||
-					( item.templatePart && item.templatePart.has_theme_file ) )
-		);
+		const isResetting = items.every( ( item ) => item?.has_theme_file );
 
 		const promiseResult = await Promise.allSettled(
 			items.map( ( item ) => {